+++ NEWS.md
@@ -6,6 +6,7 @@
    - Use https protocol for getting NCBI data
    - `dumpNcbiTax`: using the zip archive instead of the tar.gz because
    of internal issue with firewall.
+   - Correct canonical products from Ensembl
 
 <!----------------------------------------------------------------------------->
 ## Version 1.4.4